requirements.txt
Flask>=2.2.0
Flask-Cors>=3.0.9
numpy>=1.20.0
Pillow>=9.0.0
gunicorn>=20.1.0 